+++ template-svelte-js/src/sidepanel/index.js
@@ -1,0 +1,7 @@
+import App from './Sidepanel.svelte'
+
+const app = new App({
+  target: document.getElementById('app'),
+})
+
+export default app